requirements.txt
crewai
python-dotenv
crewai_tools
requests
